+++ packages/block-library/src/terms-query/edit/inspector-controls/order-control.jsx
@@ -1,9 +1,9 @@
 import { __ } from '@wordpress/i18n';
-import { SelectControl } from '@wordpress/components';
+import { SelectControl as WCSelectControl } from '@wordpress/components';
 
 export default function OrderControl( { orderBy, order, onChange, ...props } ) {
 	return (
-		<SelectControl
+		<WCSelectControl
 			options={ [
 				{
 					label: __( 'Name: A → Z' ),