+++ packages/components/src/validated-form-controls/components/stories/content-editable-control.story.tsx
@@ -38,7 +38,7 @@ export const Default: StoryObj< typeof ValidatedContentEditableControl > = {
 						? {
 								type: 'invalid',
 								message: 'The word "error" is not allowed.',
-						  }
+							}
 						: undefined
 				}
 			/>