+++ packages/ui/src/form/primitives/input-layout/stories/index.story.tsx
@@ -10,11 +10,6 @@ const meta: Meta< typeof InputLayout > = {
 	subcomponents: {
 		'InputLayout.Slot': InputLayout.Slot,
 	},
-	// Temporary: Due to an upstream bug, render the component explicitly so the
-	// components manifest extractor can resolve props from the JSX.
-	//
-	// See: https://github.com/storybookjs/storybook/issues/34877
-	render: ( args ) => <InputLayout { ...args } />,
 	parameters: {
 		componentStatus: {
 			status: 'recommended',