+++ packages/widget-dashboard/src/test/actions.jsdom.test.tsx
@@ -208,7 +208,7 @@ describe( 'WidgetDashboard.Actions', () => {
 		);
 
 		expect(
-			await screen.findByRole( 'menuitem', { name: 'Reset to default' } )
+			await screen.findByRole( 'menuitem', { name: 'Reset to default…' } )
 		).toBeInTheDocument();
 	} );